edit

+++ database/migrations/2024_04_05_141740_create_progress_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users');
             $table->integer('Current_weight')->default(0);
             $table->integer('age');
-            $table->bigInteger('points')->default(0);
+            $table->integer('points')->default(0);
             $table->integer('height');
             $table->enum('gender',['feminine','man']);
             $table->integer('Target_weight')->default(0);
